prov/psm3: Convert error message to debug output
In pmodels/mpich#6518, a user reported seeing a "No HALs registered"
message when using MPICH+libfabric. The error is because the psm3
provider is not compatible with --disable-shared. MPICH will just ignore
the error and try to pick another provider, so we convert it to debug
output to avoid spamming the user.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>